Extract changeWriterInput
